Allow installation on Nextcloud 14 (master branch)
The Nextcloud master branch is on version 14, and upon updating to it yesterday I noticed that it refused to enable the OnlyOffice app because the max-version was too low. I haven't noticed any issues with using the app on the nightly after force-installing it by making this change locally, so I'd like to propose this change in order to add compatibility with Nextcloud 14.